Software: Changed meta key to work as a modifier key
As long as the meta key is held down, the software listens for and
executes key bindings. The key is relayed on release if no bound key
was pressed in the meantime.

This avoids having to press the meta key twice in order to actually
type it. The only difference to other keys is that the character does
not appear immediately on key down but on release. This also makes
issuing successive commands much easier.