Hardware: Removed EXTVCC Jumper, removed copper around edges
There is no VCC on the keyboard connector, so VCC must be fed from external
source anyways if the keyboard cable is simply cut in half and the board
soldered in between.

The copper/traces around the keyboard/cia connectors have been removed to
make it easier to simply solder the wires in instead of adding pin headers
and proper plugs.