Software/Firmware: Simplified key encoding (numbered from 0-63)
Instead of encoding the row/columns in the high and low nibbles, we now
simply number the keys in a linear fashion corresponding to their position
in the matrix (starting from the top left at 0:0 downto 7:7).

This greatly simplifies the code and creates a direct correlation of key
symbols to the addresses used for the crosspoint switch.