Firmware: Fixed typing routine for repeated strokes of the same key
Apparently the kernal key scan routine requires at least 20ms between
successive strokes of the same key, whereas typing a different key needs
no delay between the release of the first key and the press of the second
key. In any case the key needs to be held down for 20ms as well in order
to be scanned correctly.