Firmware: Fixed serial commands for key down and key up
When a key is held down via the serial interface, it will remain down
until it is released via the serial interface up command. The respective
respective key is locked while it is down, meaning the state of the
physical keyboard key is excluded from being permanently relayed to the
crosspoint switch matrix facing the C64.