Software/Firmware: Allow configuring crosspoint switch model at runtime
This avoids having to build/ship different versions of the firmware
for the crosspoint switch IC model in use.

The configuration command 'using' has been added for this purpose:

using CD74HC22106
using MT8808

The default model is CD74HC22106 to maintain backwards compatibility.

These ICs are pin compatible but not signal-compatible: STROBE and MR
are active-low for the 22106 and active-high for the MT8808.