Hardware: Added option to power device from USB or external source
This makes it easier to use the device as a standalone keyboard controller,
i.e. without actually connecting it to a computer.

The software will support a standalone mode, in which no crosspoint switch
and consequently no meta key is required. The port used for the crosspoint
switch control will be made available as an additional user control port.