Software/Firmware: Added --delay option to specify automatic delays
The --delay option can be used to specify the time in milliseconds that
the device should wait after executing each command from a sequence of
immediate commands received via the USB control interface.

The client now also accepts a filename denoting a file that contains
the commands to send via usb. This allows writing executable keyman64
scripts using the #! mechanism.