Software: Added --preserve option for convert command
When the preserve option is used with the convert command, the currently
saved state of the control lines is fetched from the device via USB and
is then included in the newly generated binary configuration. This allows
the saved state to be preserved across reconfigurations.

If the device cannot be accessed, the state will be initiated to the default
state (all lines tristated), which is the default behaviour when --preserve
is not specified. A warning will be issued in this case.