Firmware/Software: Allow executing different commands every other time
It is now possible to specify an execution policy for commands that controls
whether a command is always executed each time a binding is triggered or
only every other time (even or odd):

<key>:[<policy>] <command>

Where policy is either "0" or "1" for even or odd. If no policy is given,
the default is to always execute the command (like before). For example:

D:0 clear port a bit 0
D:1 tristate port a bit 0

When you press <meta>-D for the first time, the clear command is executed.
The second time you press <meta>-D, the tristate command is executed, and
so on. This makes a single key act more like an actual switch rather than
a simple button.

Suggested by Thomas "GMP" Müller.