Version 1.3
Changes in this version:

Hardware:

* Updated to board revision 4
* Fixed wrong USB level conversion circuit
* Placed resistors and diodes horizontally
* Fixed wrong button foorprint
* Removed hole for pin 2 at keyboard/computer connectors

Software:

* Added custom symbol definitions for configuration files
* Added frequency test to fallback configuration (<meta>-f)

Notes:

Upgrades to the software and firmware are not mandatory.

In order to use custom symbols in configuration files, only the
client software needs to be updated.

In order to use the new frequency test in the fallback configuration, only
the firmware needs to be upgrades.