Firmware/Software: Sanitized serial interface, implemented mappings
The serial interface has been cleaned up and now accepts the command as a
byte instead of a nibble. In addition, it is now possible to define
commands expecting more than one argument byte. The example code in
serial.asm has been adjusted accordingly. This is a breaking change.

It is now also possible to map control lines to keys, where a line is
used as an active low input. As long as the line is driven low, the
mapped key is held down.

Mappings are defined by using the new map command:

map <port> <bits> [to] <key>

A corresponding serial command has been added as well:

00000101 <line> <key>

Where line contains the port in bit 3 (0 = port a, 1 = port b). This allows
changing the mapping on the fly, e.g. from the host computer.

This feature might be useful for using additional buttons
as keys, e.g. for joysticks with more than one button.