Translates between a CBM600 keyboard and the C64
This branch is a hack of version 1.4 which connects a cbm600 keyboard
to a C64 with the help of a modified scan circuit and method.

It constitutes a proof of context for an alternative scan circuit that
will allow an upcoming hardware revision to be adjusted to any
computer/keyboard combination using a matrix keyboard of up to 8
columns and 16 rows.

The code here is just a quick hack and will need to be refactored
greatly to allow compatibility with previous hardware revisions and to
allow user-configurable mapings of arbitrary keyboard/computer
combinations.

Thanks to ThunderBlade/BLiSS for requesting this hack.