Firmware: Refactored storage for password and state, deprecated --preserve
The storage method for the password and the saved control line
state has been refactored. Especially state is no longe stored
alongside the binary configuration. Instead, all transient data
is stored at the end of the eeprom memory.

Thus stored state does not need to be fetched from the device using
the --preserve option when creating a new binary configuration. The
--preserve option has been deprecated accordingly.

In addition, the password is initialized to an empty password when
no password is found in eeprom memory. The lock command now types
"no password defined" when an empty password is encountered.

When updating to this firmware, previously stored state is lost.