Firmware/Software: Implemented port expansion support
It is now possible to expand the number of control pins
by using one or more (daisy chained) 74HC595 shift
registers. Only four native control pins are required to
drive the shift registers. The new configuration directive
"expand" can be used to specify the number of additional
registers and the pins used to control the shift register
chain. Once configured, additional ports can be used like
the existing ports a and b and will be called c, d, e etc.

The additional ports can only be used as outputs and can only
be driven high or low.

The binary configuration format has been modified, so the
configuration needs to be reflashed after update.