Software: Improved firmware update procedure
The update command now accepts a configuration file as an optional
second argument. If specified, the current configuration will be
disabled before updating the firmware and the given configuration
file will be converted and installed after the firmware update.

This is intended for updates that include breaking changes to the
binary configuration format, i.e. updates that would cause the new
firmware to break when trying to read an old configuration format.