Software/Firmware: Removed device command for reset, kept client command
It is not really required to be able to bind reset to a key since
it is only useful to get the device out of bootloader mode, in which
case keybindings won't work anyways. And it is very likely that users
are already using a symbol of the same name for their reset sequence,
thus introducing a command would break existing configurations.

However, the device can still be reset from the client side using
the corresponding client command.