Firmware: Removed unnecessary assertion of crosspoint switch reset line
Asserting the reset signal of the crosspoint switch does not seem to
work for all supported models of switches, thus the reset is already
additinally done in software.

All that is left to do is initally setup the reset line, i.e. enable the
crosspoint switch. This has been moved to a new SetupCrosspointSwitch()
function for clarity.

In hindsight, driving the reset line from the avr would not have
been necessary at all if I hadn't decided to add the option of using
different crosspoint switches (CP22106 and MT8808 only differ on whether
the control signals are active low or active high).

I might change this in another revision and simply add a jumper that pulls
the reset line to GND or VCC in order to configure the crosspoint switch
model.