Software: Open usb devices using their actual device nodes in linux
Up to now the --device argument has not been used at all, instead the
device was looked up by iterating over all usb devices until a device
with a matching VID and PID was found. This made it impossible to choose
a specific device if more than one keyman64 device was connected.

Now the actual device nodes are used so that the devices can be explicitly
chosen using the --device argument.